chore: refactor common methods to utils
